Move integration type and interaction contexts
